Modification to proposal and sample notebook flow
After the proposal is rejected, user will again submit the proposal.
So the proposal form is filled with the initial values of the rejected
proposal.

Sample notebook can be submitted only if the proposal is approved else
will display a message saying 'not permitted'

User after sample notebook submission can anytime edit it.

Modified the template for the reviewer to see the new and old proposal.
Since the ids where same for the collapse element in the loop so there
was problem in javascript.

Profile view modified.
First checks for the user profile. If exists then updates or else
creates a new profile for the user. This will prevent multiple profile
for a single user.[Alternatively, we can give one-to-one relation between
user and profile in models]